// WebSystem/WebSystem/ClientApp/src/models/Formas/CNT/dpnumero.ts
export interface IDpnumero {
  id?: number;
  tipo_asi: string;  
  nombre?: string;  
  asiento?: string;  
  nomforma?: string;  
  tipo_com?: string;  
  modulo?: string;  
  ch_boucher?: string;  
  tiene_at?: string;  
  secumes?: string;  
  serie_rt?: string;  
  autori_rt?: string;  
  celectro?: string;  
  es_retenci?: string;  
  fiscal?: boolean;
}